renamed initialise_solutions to initialising
